Search functions
Added functions to search adjacent to an element in the tree (or where an element should be, if it were in the tree). Also exposed some convenience functions in for getting the root data and making a search query into the tree to see if something exits, returning the data if successful.

Corrected a typo in the readme and updated header definition shown.